fix: separation of concerns on controllers; feat: new controllers include file
