+++ app/controllers/rolController.php
@@ -3,37 +3,43 @@ require_once "app/models/RolModel.php";
 require_once "app/models/PermisoModel.php";
 require_once "app/models/AuthModel.php";
 
-class RolController{
+class RolController
+{
     private $rolModel;
     private $permisoModel;
 
-    public function __construct(){
+    public function __construct()
+    {
         $this->rolModel = new Rol();
         $this->permisoModel = new Permiso();
     }
-
-    public function fetchPermisos(){
-        $roles=$this->permisoModel->get_permiso();
+    public function fetchRoles()
+    {
+        $roles = $this->rolModel->getRoles();
         header("Content-Type: application/json");
         echo json_encode($roles);
+
         exit();
     }
 
-    public function index(){
+
+    public function index()
+    {
         require "app/views/roles.php";
     }
 
-    public function crearRol() {
+    public function crearRol()
+    {
         header('Content-Type: application/json');
         $nombre_rol = trim($_POST['nombre_rol']);
         $descripcion_rol = trim($_POST['descripcion_rol']);
         $permisos = isset($_POST['permisos']) ? $_POST['permisos'] : [];
         $result = $this->rolModel->createRol($nombre_rol, $descripcion_rol);
-    
+
         if ($result) {
             // Obtener el ID del nuevo rol
             $idNuevo = $this->rolModel->getId($nombre_rol);
-            
+
             if ($idNuevo) {
                 // Asociar los permisos seleccionados
                 $autoriza = new Autorizacion();
