+++ router.php
@@ -9,25 +9,7 @@ if (isset($_SESSION["user_id"])) {
     Model::setUsuarioActivo($_SESSION["user_id"]);
 }
 
-require_once "app/controllers/sessionController.php";
-require_once "app/controllers/userController.php";
-require_once "app/controllers/dashboardController.php";
-require_once "app/controllers/profileController.php";
-require_once "app/controllers/incomeController.php";
-require_once "app/controllers/rolController.php";
-require_once "app/controllers/expenseController.php";
-require_once "app/controllers/reportController.php";
-require_once "app/controllers/auditController.php";
-
-$incomeController = new IncomeController();
-$profileController = new ProfileController();
-$sessionController = new SessionController();
-$userController = new UsuarioController();
-$dashboardController = new DashboardController();
-$rolController = new RolController();
-$expenseController = new ExpenseController();
-$reportController = new ReportController();
-$auditController = new AuditController();
+require("config/includes/controllers.php");
 
 $page = isset($_GET["page"]) ? $_GET["page"] : "dashboard";
 $route = isset($_GET["route"]) ? $_GET["route"] : "";
@@ -48,11 +30,11 @@ if (isset($_GET["route"])) {
             break;
 
         case "get-permisos":
-            $rolController->fetchPermisos();
+            $permitController->fetchPermisos();
             break;
         
         case "get-roles":
-            $userController->fetchRoles();
+            $rolController->fetchRoles();
             break;
 
         case "crear-rol":
